+++ src/ts/main/main-folder.ts
@@ -1,4 +1,4 @@
-import { app, ipcMain } from "electron";
+import { ipcMain } from "electron";
 import * as path from "path";
 import * as fs from "fs";