+++ src/ts/renderer/app.ts
@@ -8,6 +8,8 @@ import * as Bridge from "./utils/utils.bridge.js";
 import * as Requests from "./utils/utils.requests.js";
 import "./utils/utils.types.js";
 
+const w: any = (window as any);
+
 export default class App {
     private window: Window;
     public account: Account;
@@ -32,7 +34,7 @@ export default class App {
     }
 
     private initEvents(): void {
-        (window as any).electronAPI.onBeforeClose(async () => await this.saveSettings());
+        w.mainEvents.onClose(async () => await this.saveSettings());
     }
 
     public async init(): Promise<void> {